moved sendMessage and onMounted with socket.on(messageAdded) and input reactive state to ChatBox
