+++ frontend/src/views/ChatView.vue
@@ -19,21 +19,10 @@
             </div>
             <div class="col-12 md:col-8 xl:col-9">
               <ChatBox />
-              <div class="card">
+              <!-- <div class="card">
                 <form class="form" @submit.prevent="onSubmit">
-                  <input
-                    v-model="input"
-                    placeholder="Type your message..."
-                    class="input"
-                    @keyup.enter="sendMessage"
-                  />
-                  <div class="input-group-append">
-                    <button class="send-button" @click="sendMessage">
-                      Send
-                    </button>
-                  </div>
                 </form>
-              </div>
+              </div> -->
             </div>
           </div>
         </div>
@@ -44,7 +33,7 @@
 </template>
 
 <script setup lang="ts">
-import { onMounted, ref, inject } from "vue";
+import { onMounted, inject } from "vue";
 import { useRouter } from "vue-router";
 import { Socket } from "socket.io-client";
 
@@ -52,7 +41,6 @@ import ChatRoomsList from "../components/ChatRoomsList.vue"; //TODO: update tsco
 import ChatBox from "../components/ChatBox.vue"; // @ is an alias to /src
 import CreateRoom from "./CreateRoom.vue"; // Vetur still doesn't support script setup so is unhappy
 // import SocketioService from "../services/socketio.service";
-import MessageI from "../types/Message.interface";
 import PrimeVueButton from "primevue/button";
 
 const router = useRouter();
@@ -66,20 +54,7 @@ onMounted(() => {
   socket.on("messages", (args: string) => {
     args;
   }); //to get all messages of the user for this room?
-
-  socket.on("messageAdded", (message: MessageI) => {
-    console.log("messageAdded event received from backend");
-    console.log("Here is you msg saved to DB: ", message.text);
-  }); //event triggered when a msg is saved to DB
 });
-
-// reactive input state
-const input = ref<string>("");
-//binding a click event listener to a method named 'sendMessage'
-function sendMessage() {
-  console.log("input value: ", input.value);
-  socket.emit("addMessage", { text: input.value });
-}
 
 function openCreateRoomCard() {
   router.push({
